Move server functions from cookie_consent.rs to server.rs
